Files moved, added calculate_pixmap
